V1.3 - GUI Working - Support for 3 Screens
GUI Works perfectly, HTML might come later or as a different project. I didn't really use it in this scenario so fixing it seemed redundant.

The config will reset if the amount of devices changes, and the drop-down option "Select Device x" disappears after a selection is made.

I will append the page with a correct readme later (:

Dependancies are XRANDR & pyqt6